refactor(backoffice): share script presentation components
Extract the code viewer and presentation controls from the automation
source panel. Allow canonical source and graph views to fill bounded
containers for reuse by session workspaces.
