+++ apps/backoffice/app/routes/backoffice/automations/script-view/script-source-panel.tsx
@@ -1,5 +1,4 @@
-import { Code2, Columns2, Workflow as WorkflowIcon } from "lucide-react";
-import { useEffect, useMemo, useRef, useState, type ComponentType } from "react";
+import { useMemo, useState } from "react";
 import { useSearchParams } from "react-router";
 
 import { visualizeWorkflowSource, type SourceRange } from "@fragno-dev/workflow-visualizer-tokens";
@@ -10,7 +9,9 @@ import {
   type RuntimeToolWorkflowDescriptor,
 } from "@/fragno/runtime-tools/workflow-catalog";
 
-import { useLinkedScrollViewports, type LinkedScrollViewport } from "./linked-scroll";
+import { useLinkedScrollViewports } from "./linked-scroll";
+import { ScriptCodeView } from "./script-code-view";
+import { ScriptViewToggle, WorkflowGraphDetailToggle } from "./script-presentation-controls";
 import {
   SCRIPT_VIEW_MODE_SEARCH_PARAM,
   WORKFLOW_GRAPH_DETAIL_MODE_SEARCH_PARAM,
@@ -19,22 +20,10 @@ import {
   searchParamsWithScriptViewMode,
   searchParamsWithWorkflowGraphDetailMode,
   workflowGraphDetailModeFromSearchParam,
-  type ScriptViewMode,
-  type WorkflowGraphDetailMode,
 } from "./script-view-mode";
 import { useScriptWorkflowRuns } from "./use-script-workflow-runs";
 import { ScriptWorkflowGraph } from "./workflow-graph";
 import type { ScriptWorkflowRun } from "./workflow-run-presentation";
-
-const SCRIPT_VIEW_OPTIONS: Array<{
-  mode: ScriptViewMode;
-  label: string;
-  icon: ComponentType<{ className?: string }>;
-}> = [
-  { mode: "code", label: "Code", icon: Code2 },
-  { mode: "graph", label: "Graph", icon: WorkflowIcon },
-  { mode: "split", label: "Both", icon: Columns2 },
-];
 
 export function ScriptSourcePanel({
   absolutePath,
@@ -222,135 +211,5 @@ function WorkflowRunSelector({
         ))}
       </select>
     </label>
-  );
-}
-
-function ScriptViewToggle({
-  viewMode,
-  onViewModeChange,
-}: {
-  viewMode: ScriptViewMode;
-  onViewModeChange: (mode: ScriptViewMode) => void;
-}) {
-  return (
-    <div
-      role="group"
-      aria-label="Script view"
-      className="flex shrink-0 border border-[color:var(--bo-border)] bg-[var(--bo-panel-2)] p-0.5"
-    >
-      {SCRIPT_VIEW_OPTIONS.map(({ mode, label, icon: Icon }) => (
-        <button
-          key={mode}
-          type="button"
-          aria-pressed={viewMode === mode}
-          onClick={() => {
-            onViewModeChange(mode);
-          }}
-          className={`${segmentedToggleButtonClass(viewMode === mode)} gap-1.5`}
-        >
-          <Icon className="h-3.5 w-3.5" />
-          {label}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-function WorkflowGraphDetailToggle({
-  detailMode,
-  onDetailModeChange,
-}: {
-  detailMode: WorkflowGraphDetailMode;
-  onDetailModeChange: (mode: WorkflowGraphDetailMode) => void;
-}) {
-  return (
-    <div
-      role="group"
-      aria-label="Workflow graph detail"
-      className="flex shrink-0 border border-[color:var(--bo-border)] bg-[var(--bo-panel-2)] p-0.5"
-    >
-      {(["simple", "verbose"] as const).map((mode) => (
-        <button
-          key={mode}
-          type="button"
-          aria-pressed={detailMode === mode}
-          onClick={() => {
-            onDetailModeChange(mode);
-          }}
-          className={segmentedToggleButtonClass(detailMode === mode)}
-        >
-          {mode}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-function segmentedToggleButtonClass(isSelected: boolean): string {
-  const interaction =
-    "flex min-h-10 items-center px-2.5 text-[10px] font-semibold tracking-[0.12em] uppercase transition-[color,background-color,box-shadow,transform] active:scale-[0.96]";
-  return isSelected
-    ? `${interaction} bg-[var(--bo-panel)] text-[var(--bo-fg)] shadow-sm`
-    : `${interaction} text-[var(--bo-muted-2)] hover:text-[var(--bo-fg)]`;
-}
-
-function ScriptCodeView({
-  script,
-  split,
-  selectedSource,
-  scrollViewport,
-  onSourceReveal,
-}: {
-  script: string;
-  split: boolean;
-  selectedSource?: SourceRange;
-  scrollViewport: LinkedScrollViewport;
-  onSourceReveal: () => void;
-}) {
-  const selectionRef = useRef<HTMLElement>(null);
-  const selectionStart = Math.max(0, Math.min(script.length, selectedSource?.start.offset ?? 0));
-  const selectionEnd = Math.max(
-    selectionStart,
-    Math.min(script.length, selectedSource?.end.offset ?? selectionStart),
-  );
-  const hasSelection = selectionEnd > selectionStart;
-
-  useEffect(() => {
-    if (hasSelection) {
-      onSourceReveal();
-      selectionRef.current?.scrollIntoView({ block: "center", behavior: "smooth" });
-    }
-  }, [hasSelection, onSourceReveal, selectionStart, selectionEnd]);
-
-  return (
-    <div
-      {...scrollViewport}
-      tabIndex={0}
-      aria-label="Script source"
-      className={`backoffice-scroll max-h-[calc(100vh-10rem)] min-h-[36rem] overflow-auto focus-visible:outline-2 focus-visible:-outline-offset-2 focus-visible:outline-[color:var(--bo-accent)] ${split ? "border-b border-[color:var(--bo-border)] lg:border-r lg:border-b-0" : ""}`}
-    >
-      <pre className="min-h-full px-4 py-4 font-mono text-xs break-words whitespace-pre-wrap text-[var(--bo-fg)]">
-        <code>
-          {script ? (
-            hasSelection ? (
-              <>
-                {script.slice(0, selectionStart)}
-                <mark
-                  ref={selectionRef}
-                  className="bg-amber-300/35 text-inherit outline outline-1 outline-amber-500/50 dark:bg-amber-300/20"
-                >
-                  {script.slice(selectionStart, selectionEnd)}
-                </mark>
-                {script.slice(selectionEnd)}
-              </>
-            ) : (
-              script
-            )
-          ) : (
-            "# Empty script"
-          )}
-        </code>
-      </pre>
-    </div>
   );
 }
